[WIP] Add a watchdog timer to measurement
+ It is a timer instance that has a longer delay than measurement time.
If the measurement succeeds before it triggers, it'll be aborted
automatically. However when if it triggers, it is treated as an error
that something is lost during measurement.

+++ projects/depscan-firmware-rtos/Src-Usr/app/dist-sensor.h
@@ -32,7 +32,7 @@ typedef struct dist_sens__* dist_sens_t;
 //! @brief      Sensor configuration descriptor
 typedef struct dist_sens_config__
 {
-    uint16_t Delay_us;           //!< Delays per capture
+    uint32_t Delay_us;           //!< Delays per capture
     bool     bCloseDistanceMode; //!< Distance mode. True if close distance &
                                  //!< accuracy mode
 } dist_sens_config_t;
